requirements.txt
Django==2.1.7
ipython==6.5.0
psycopg2==2.7.7
django_heroku==0.3.1
gunicorn==19.9.0
Pillow==5.2.0
whitenoise==4.1
django-cors-headers==2.5.2
djangorestframework==3.9.2
